Break out hvac_integration_test_case
