+++ tests/integration_tests/api/auth_methods/test_azure.py
@@ -6,10 +6,11 @@ from parameterized import parameterized, param
 
 from hvac import exceptions
 from tests import utils
+from tests.utils.hvac_integration_test_case import HvacIntegrationTestCase
 
 
 @skipIf(utils.skip_if_vault_version_lt('0.10.0'), "Azure auth method not available before Vault version 0.10.0")
-class TestAzure(utils.HvacIntegrationTestCase, TestCase):
+class TestAzure(HvacIntegrationTestCase, TestCase):
     TEST_MOUNT_POINT = 'azure-test'
 
     def setUp(self):
