+++ tests/integration_tests/api/auth_methods/test_gcp.py
@@ -6,9 +6,10 @@ from parameterized import parameterized, param
 
 from hvac import exceptions
 from tests import utils
+from tests.utils.hvac_integration_test_case import HvacIntegrationTestCase
 
 
-class TestGcp(utils.HvacIntegrationTestCase, TestCase):
+class TestGcp(HvacIntegrationTestCase, TestCase):
     TEST_MOUNT_POINT = 'gcp-test'
 
     def setUp(self):
