+++ tests/integration_tests/api/system_backend/test_key.py
@@ -1,10 +1,12 @@
 import logging
 from unittest import TestCase
+
 from hvac import exceptions
 from tests import utils
+from tests.utils.hvac_integration_test_case import HvacIntegrationTestCase
 
 
-class TestKey(utils.HvacIntegrationTestCase, TestCase):
+class TestKey(HvacIntegrationTestCase, TestCase):
 
     def test_start_generate_root_with_completion(self):
         test_otp = 'RSMGkAqBH5WnVLrDTbZ+UQ=='
