+++ tests/integration_tests/api/system_backend/test_policy.py
@@ -2,10 +2,11 @@ from unittest import TestCase
 from unittest import skipIf
 
 from tests import utils
+from tests.utils.hvac_integration_test_case import HvacIntegrationTestCase
 
 
 @skipIf(utils.skip_if_vault_version_lt('0.9.0'), "Policy class uses new parameters added >= Vault 0.9.0")
-class TestPolicy(utils.HvacIntegrationTestCase, TestCase):
+class TestPolicy(HvacIntegrationTestCase, TestCase):
 
     def test_policy_manipulation(self):
         self.assertIn(
