+++ tests/integration_tests/v1/test_approle.py
@@ -4,10 +4,10 @@ from unittest import TestCase
 from parameterized import parameterized, param
 
 from hvac import exceptions
-from tests import utils
+from tests.utils.hvac_integration_test_case import HvacIntegrationTestCase
 
 
-class TestApprole(utils.HvacIntegrationTestCase, TestCase):
+class TestApprole(HvacIntegrationTestCase, TestCase):
     TEST_MOUNT_POINT = 'approle'
 
     def setUp(self):
