+++ tests/integration_tests/v1/test_system_backend.py
@@ -5,9 +5,10 @@ from parameterized import parameterized, param
 
 from hvac import exceptions
 from tests import utils
+from tests.utils.hvac_integration_test_case import HvacIntegrationTestCase
 
 
-class TestSystemBackend(utils.HvacIntegrationTestCase, TestCase):
+class TestSystemBackend(HvacIntegrationTestCase, TestCase):
 
     def test_unseal_multi(self):
         cls = type(self)
