+++ tests/unit_tests/api/secrets_engines/test_kv.py
@@ -6,10 +6,9 @@ from parameterized import parameterized
 from hvac.api.secrets_engines.kv import Kv
 from hvac.api.secrets_engines.kv_v1 import KvV1
 from hvac.api.secrets_engines.kv_v2 import KvV2
-from tests import utils
 
 
-class TestKv(utils.HvacIntegrationTestCase, TestCase):
+class TestKv(TestCase):
 
     def test_v1_property(self):
         mock_adapter = MagicMock()
